+++ components/Providers.tsx
@@ -6,6 +6,7 @@ import { ConsentProvider } from './gdpr/ConsentProvider'
 import CookieBanner from './gdpr/CookieBanner'
 import CookiePreferencesModal from './gdpr/CookiePreferencesModal'
 import ConsentScripts from './gdpr/ConsentScripts'
+import FeedbackWidget from './shared/FeedbackWidget'
 
 export default function Providers({ children }: { children: React.ReactNode }) {
   return (
@@ -13,6 +14,7 @@ export default function Providers({ children }: { children: React.ReactNode }) {
       <ConsentProvider>
         <ConsentScripts />
         {children}
+        <FeedbackWidget />
         <SurveyModalV2 />
         <CookieBanner />
         <CookiePreferencesModal />